refactoring the logic for game objects
