+++ src/window/Window.c
@@ -1,7 +1,6 @@
 #include "Window.h"
 
-
-#include "math/mathUtils.h"
+#include "game/GameObjects/Cube.h"
 
 int WINDOW_WIDTH = 1280;
 int WINDOW_HEIGHT = 750;
@@ -10,8 +9,7 @@ int frame = 0;
 int timebase = 0;
 int fps = 0;
 
-Time time =
-{
+Time time = {
 	.currTime = 0,
 	.prevTime = 0.0f,
 	.deltaTime = 0.0f,
@@ -19,29 +17,6 @@ Time time =
 
 
 GameObjectManager gameObjectManager;
-
-
-void onCubeUpdate(Time time, GameObject* gameObject)
-{
-	Vector3* pos = &gameObject->transform.position;
-
-	float amplitude = 2.0f;
-	float period = 2.0f;
-
-	float theta = (time.currTime) * period;
-	float distance = sin(theta) / amplitude;
-
-	float hover = lerp(6.0f, 12.0f, distance);
-
-	pos->y = hover;
-
-
-	Vector3* rot = &gameObject->transform.rotation;
-	rot->x += (time.deltaTime * 20);
-	rot->y += (time.deltaTime * 20);
-	rot->z += (time.deltaTime * 20);
-
-}
 
 
 void initialiseWindow(int* argc, char** argv, char* windowName)
@@ -96,78 +71,11 @@ void initialiseWindow(int* argc, char** argv, char* windowName)
 	// setup game object manager
 	initGameObjectManager(&gameObjectManager);
 
-	GameObject cube;
 	initGameObject(&cube);
-
-	const Vector3 cubeVertexBuffer[] = {
-		{ -1.0f, -1.0f, -1.0f },
-		{ 1.0f, -1.0f, -1.0f },
-		{ -1.0f,  1.0f, -1.0f },
-		{ 1.0f,  1.0f, -1.0f },
-
-
-		{ 1.0f,  1.0f,  1.0f },
-		{ 1.0f, -1.0f,  1.0f },
-		{ -1.0f,  1.0f,  1.0f },
-		{ -1.0f, -1.0f,  1.0f },
-	};
-
-	const Vector3Int cubeIndexBuffer[] = {
-		// front
-		{0, 2, 1},
-		{1, 3, 2},
-		// right
-		{1, 3, 4},
-		{4, 5, 1},
-		// back
-		{5, 4, 6},
-		{6, 7, 5},
-		// left
-		{7, 6, 0},
-		{0, 2, 6},
-		// bottom
-		{0, 1, 7},
-		{7, 5, 1},
-		// top
-		{2, 3, 6},
-		{6, 4, 3}
-	};
-
-	const RGBA cubeColorBuffer[] = {
-		// front colors
-		{1.0f, 0.0f, 0.0f, 1.0f},
-		{0.0f, 1.0f, 0.0f, 1.0f},
-		{0.0f, 0.0f, 1.0f, 1.0f},
-		{1.0f, 1.0f, 1.0f, 1.0f},
-		// back colors
-		{1.0f, 0.0f, 0.0f, 1.0f},
-		{0.0f, 1.0f, 0.0f, 1.0f},
-		{0.0f, 0.0f, 1.0f, 1.0f},
-		{1.0f, 1.0f, 1.0f, 1.0f}
-	};
-
-	Mesh cubeMesh = {
-		.points = cubeVertexBuffer,
-		.indices = cubeIndexBuffer,
-		.pointSize = 8,
-		.indexCount = 36,
-		.colors = cubeColorBuffer,
-		.isUniformColor = false,
-		.debug = false
-	};
-
-	cube.name = "cube";
-
-	cube.mesh = cubeMesh;
-	calculateMeshBoundBox(&cube.mesh);
-
-	cube.transform.position = (Vector3){ 5.0f, 4.0f, 0.0f };
-	cube.transform.scale = (Vector3){ 1.0f, 1.0f, 1.0f };
-
-	cube.onUpdate = &onCubeUpdate;
-
 	gameObjectManagerAdd(&gameObjectManager, cube);
 
+
+	gameObjectManager.gameObjects[0].onStart(&gameObjectManager.gameObjects[0]);
 
 	GameObject floor;
 	initGameObject(&floor);
